predictor/requirements.txt
grpcio
grpcio-tools
google-api-core
googleapis-common-protos
grpc-gateway-protoc-gen-openapiv2
protobuf == 3.20.*
torch
scipy
numpy
attrs
openpyxl
pymongo
python-dotenv
sentence_transformers == 3.0.0
pandas == 2.2.0
scikit-learn == 1.5.0
colbert-ai 
faiss-cpu
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation
opentelemetry-semantic-conventions
opentelemetry-exporter-otlp